WIP Add support for Unicode Windows
Adding support for unicode characters and emoticons without the use of wincompose.
There are two built in ways for windows 10 to support unicode: ALTX and ALTPLUS
ALTX is contextually blased and requires being in an app that supports Windows rich text like Word and Outlook.
ALTPLUS allows for unicode to be entered in any application.

ALTX is currently fully implemented
ALTPLUS is currently a WIP and being tested.